Changes for dashboard page completed in topics components. They will be renamed in feauture PR. Footer 'topics' menu has removed from footer.
