--- layout/app/dashboards-detail/index.js
@@ -1,10 +1,0 @@
-// Redux
-import { connect } from 'react-redux';
-
-// component
-import DashboardDetailLayout from './component';
-
-export default connect(
-  state => ({ topicsDetail: state.topics.detail.data }),
-  null
-)(DashboardDetailLayout);
