--- layout/dashboards/index.js
@@ -1,8 +1,0 @@
-import { connect } from 'react-redux';
-
-import DashboardsLayout from './component';
-
-export default connect(
-  state => ({ data: state.staticPages.topics }),
-  null
-)(DashboardsLayout);
